# docs/openapi.yml
openapi: 3.0.3
info:
  title: Tournament Platform API
  version: v1
  description: |
    API for the Tournament Platform, enabling user management, wallet operations, tournament participation, and administrative functions.
    Key features include secure authentication, real-time match updates, and role-based access control.
  contact:
    name: Support Team
    email: support@example-tournament.com
  license:
    name: ISC
    url: https://opensource.org/licenses/ISC
servers:
  - url: http://localhost:{port}/api/v1 # Define port in variables or replace directly
    description: Local development server
    variables:
      port:
        default: '3000'
        description: Port for the local server
  - url: https://staging.example-tournament.com/api/v1
    description: Staging server
  - url: https://api.example-tournament.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access Token for authentication. Example: 'Bearer {token}'"
    refreshTokenCookie: # For documenting the refresh token mechanism
      type: apiKey
      in: cookie
      name: jid # Matches JWT_REFRESH_COOKIE_NAME from .env.example
      description: "HttpOnly cookie containing the Refresh Token."
  schemas:
...
    UserPublicProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-1234-567890abcdef"
        username:
          type: string
          example: "player123"
        role:
          type: string
          enum: [User, Admin, DisputeModerator, FinanceManager]
          example: "User"

    # Wallet Schemas
...
    TournamentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        gameName:
          type: string
        status:
          type: string
        entryFee:
          type: number
          format: float
        prizePool:
          type: number
          format: float
        maxParticipants:
          type: integer
        currentParticipants:
          type: integer
        startDate:
          type: string
          format: date-time
    PaginatedTournaments:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalItems:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/TournamentSummary'
    TournamentDetails:
      allOf:
        - $ref: '#/components/schemas/TournamentSummary'
        - type: object
          properties:
            description:
              type: string
              nullable: true
            rules:
              type: string
              nullable: true
            endDate:
              type: string
              format: date-time
              nullable: true
            organizerId:
              type: string
              format: uuid
              nullable: true
            # participants: # Could be a list of participant profiles or IDs
            #   type: array
            #   items:
            #     $ref: '#/components/schemas/UserPublicProfile'
    TournamentRegistrationResponse:
      type: object
      properties:
        message:
          type: string
          example: "Successfully registered for the tournament."
        tournamentId:
          type: string
          format: uuid
        participantId: # ID of the TournamentParticipant record
          type: string
          format: uuid

    # Match Schemas
...
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    TournamentIdPath:
      name: id
      in: path
      required: true
      description: ID of the tournament.
      schema:
        type: string
        format: uuid
...
paths:
...
  /tournaments:
    get:
      tags: [Tournaments]
      summary: List tournaments with pagination and filtering.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [PENDING, REGISTRATION_OPEN, REGISTRATION_CLOSED, ONGOING, COMPLETED, CANCELED]
          description: Filter by tournament status.
        - name: gameName
          in: query
          required: false
          schema:
            type: string
          description: Filter by game name.
      responses:
        '200':
          description: A paginated list of tournaments.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTournaments'
        '500':
          description: Internal server error.
  /tournaments/{id}:
    get:
      tags: [Tournaments]
      summary: Get details of a specific tournament.
      parameters:
        - $ref: '#/components/parameters/TournamentIdPath'
      responses:
        '200':
          description: Tournament details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentDetails'
        '404':
          description: Tournament not found.
        '500':
          description: Internal server error.
  /tournaments/{id}/register:
    post:
      tags: [Tournaments]
      summary: Register the logged-in user for a tournament.
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TournamentIdPath'
      responses:
        '200':
          description: Successfully registered for the tournament.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TournamentRegistrationResponse'
        '400':
          description: Validation error (e.g., registration closed, already registered, insufficient balance for entry fee).
        '401':
          description: Unauthorized.
        '404':
          description: Tournament not found.
        '500':
          description: Internal server error.
...